# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Está API gerencia detalhes de produtos.",
        "title": "Product Details API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/compare": {
            "get": {
                "description": "Retorna uma lista de produtos com base nos IDs fornecidos em um formato de lista separada por vírgulas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Compara múltiplos produtos por uma lista de IDs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IDs dos produtos a serem comparados (ex: 0d605862-91e8-11f0-9140-00155d6d572f,034aeff8-90ea-11f0-95f2-00155d6d5ec0,034afa35-90ea-11f0-95f2-00155d6d5ec0,034b11f9-90ea-11f0-95f2-00155d6d5ec0)",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ResponseManyPaginated"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/products": {
            "post": {
                "description": "Cria um novo produto com os dados fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Cria um novo produto",
                "parameters": [
                    {
                        "description": "Dados do produto a ser criado",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Erro na requisição",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "delete": {
                "description": "Deleta um produto específico com base no ID fornecido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Deleta um produto pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do produto a ser deletado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado com sucesso\"}",
                        "schema": {
                            "$ref": "#/definitions/dto.OutputDefault"
                        }
                    },
                    "404": {
                        "description": "produto não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Retorna uma lista de produtos, paginada",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Lista todos os produtos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Numero da página a ser retornada",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade de itens na página a ser retornada",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ResponseManyPaginated"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Retorna um produto específico com base no ID fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Retorna um produto pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.OutputDefault"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza um produto existente com base no ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Atualiza um produto pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do produto para atualização",
                        "name": "produto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.OutputDefault"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.OutputDefault": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "dto.Request": {
            "type": "object",
            "properties": {
                "classificacao": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "especificacao": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                },
                "url_imagem": {
                    "type": "string"
                }
            }
        },
        "dto.Response": {
            "type": "object",
            "properties": {
                "classificacao": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "especificacao": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "url_imagem": {
                    "type": "string"
                }
            }
        },
        "dto.ResponseManyPaginated": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "itemsPerPage": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Response"
                    }
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    }
}